+++ CHANGELOG.md
@@ -5,13 +5,14 @@ All notable changes to this project will be documented in this file.
 The Ideon project follows the [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) format
 and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.7.1] - 2026-03-XX
+## [0.7.1] - 2026-03-16
 
 ### Fixed
 
 - Fixed an issue where the markdown bubble menu would not appear when expected.
 - Fixed UUID generation in non secure contexts by replacing "crypto.randomUUID" with "uuidv4"
 - Fixed Vercel integration to always keep PAT enabled, with optional OAuth.
+- Fixed touch gestures experience on mobile and touchscreen devices.
 
 ## [0.7.0] - 2026-03-16